move icons folder from folder "children" to "map", move "App.css" to new folder css, remove console.log, add boat svg as background-image for legend, change class names for legend i-elements
