+++ frontend/src/components/map/Legend.js
@@ -9,9 +9,9 @@ export default function Legend({map}) {
             legend.onAdd = () => {
                 const div = Leaflet.DomUtil.create("div", "map-legend");
                 div.innerHTML =
-                    "<p><i class='lime'></i> Route</p>" +
-                    "<p><i class='red'></i> NAVTEX</p>" +
-                    "<p><i class='blue'></i> Current position</p>";
+                    "<p><i class='route'></i> Route</p>" +
+                    "<p><i class='warning'></i> NAVTEX</p>" +
+                    "<p><i class='current-position'></i> Current position</p>";
                 return div;
             };
