+++ frontend/src/components/map/icons/MarkerIcon.js
